requirements.txt
Flask
pandas
requests
cohere
python-dotenv
